+++ js/widgets/behaviors/link-selection.ts
@@ -2,8 +2,6 @@
  * Copyright (c) 2023 ipyforcegraph contributors.
  * Distributed under the terms of the Modified BSD License.
  */
-import { GraphData } from 'force-graph';
-
 import { IBackboneModelOptions } from '@jupyter-widgets/base';
 
 import {
@@ -11,7 +9,6 @@ import {
   DEFAULT_CURVATURES,
   DEFAULT_LINE_DASHES,
   DEFAULT_WIDTHS,
-  EUpdate,
   IBehave,
   IExtraColumns,
   ILinkBehaveOptions,
@@ -50,22 +47,7 @@ export class LinkSelectionModel extends BehaviorModel implements IBehave {
   }
 
   onValueChange(change?: any) {
-    this._updateRequested.emit(EUpdate.Render);
-    if (this.columnName) {
-      this._graphDataUpdateRequested.emit(void 0);
-    }
-  }
-
-  async updateGraphData(graphData: GraphData): Promise<void> {
-    const { selected, columnName } = this;
-    if (!columnName) {
-      return;
-    }
-    const { links } = graphData;
-    const linkCount = links.length;
-    for (let i = 0; i < linkCount; i++) {
-      links[i][columnName] = selected.has(i);
-    }
+    this._updateRequested.emit(void 0);
   }
 
   get selected(): TSelectedSet {