+++ js/widgets/behaviors/node-selection.ts
@@ -2,13 +2,10 @@
  * Copyright (c) 2023 ipyforcegraph contributors.
  * Distributed under the terms of the Modified BSD License.
  */
-import { GraphData } from 'force-graph';
-
 import { IBackboneModelOptions } from '@jupyter-widgets/base';
 
 import {
   DEFAULT_COLORS,
-  EUpdate,
   IBehave,
   IExtraColumns,
   INodeBehaveOptions,
@@ -44,22 +41,7 @@ export class NodeSelectionModel extends BehaviorModel implements IBehave {
   }
 
   onValueChange(change?: any) {
-    this._updateRequested.emit(EUpdate.Render);
-    if (this.columnName) {
-      this._graphDataUpdateRequested.emit(void 0);
-    }
-  }
-
-  async updateGraphData(graphData: GraphData): Promise<void> {
-    const { selected, columnName } = this;
-    if (!columnName) {
-      return;
-    }
-    const { nodes } = graphData;
-    const nodeCount = nodes.length;
-    for (let i = 0; i < nodeCount; i++) {
-      nodes[i][columnName] = selected.has(i);
-    }
+    this._updateRequested.emit(void 0);
   }
 
   get selected(): TSelectedSet {